+++ courses/test/test_serializers.py
@@ -1,13 +1,5 @@
 import pytest
 
-from courses.models import (
-    Assignment,
-    Course,
-    Lecture,
-    MultipleChoiceQuestion,
-    MultipleChoiceQuestionChoice,
-    Topic,
-)
 from courses.serializers import (
     AssignmentSerializer,
     CourseDetailSerializer,
@@ -17,110 +9,7 @@ from courses.serializers import (
     MultipleChoiceQuestionSerializer,
     TopicSerializer,
 )
-
-# 테스트에서 사용할 상수를 정의합니다.
-COURSE_TITLE = "Test Course"
-COURSE_SHORT_DESCRIPTION = "Test Course"
-COURSE_DESCRIPTION = {}
-COURSE_CATEGORY = "JavaScript"
-COURSE_COURSE_LEVEL = "beginner"
-LECTURE1_TITLE = "Test Lecture 1"
-LECTURE1_ORDER = 1
-LECTURE2_TITLE = "Test Lecture 2"
-LECTURE2_ORDER = 2
-TOPIC1_TITLE = "Test Topic 1"
-TOPIC1_TYPE = "quiz"
-TOPIC1_DESCRIPTION = "Test Description"
-TOPIC1_ORDER = 1
-TOPIC1_IS_PREMIUM = True
-TOPIC2_TITLE = "Test Topic 2"
-TOPIC2_TYPE = "assignment"
-TOPIC2_DESCRIPTION = "Test Description"
-TOPIC2_ORDER = 2
-TOPIC2_IS_PREMIUM = True
-ASSIGNMENT_QUESTION = "Test Assignment"
-MCQ_QUESTION = "Test Multiple Choice Question"
-MCQ_CHOICE1 = "Choice 1"
-MCQ_CHOICE2 = "Choice 2"
-MCQ_CHOICE3 = "Choice 3"
-MCQ_CHOICE4 = "Choice 4"
-
-
-@pytest.fixture(autouse=True)
-def setup_course_data():
-    """
-    테스트에서 사용할 Course, Lecture, Topic, Assignment, MultipleChoiceQuestion, MultipleChoiceQuestionChoice 인스턴스를 생성합니다.
-    """
-
-    course = Course.objects.create(
-        title=COURSE_TITLE,
-        short_description=COURSE_SHORT_DESCRIPTION,
-        description=COURSE_DESCRIPTION,
-        category=COURSE_CATEGORY,
-        course_level=COURSE_COURSE_LEVEL,
-    )
-    lecture1 = Lecture.objects.create(
-        title=LECTURE1_TITLE,
-        course=course,
-        order=LECTURE1_ORDER,
-    )
-    lecture2 = Lecture.objects.create(
-        title=LECTURE2_TITLE, course=course, order=LECTURE2_ORDER
-    )
-    topic1 = Topic.objects.create(
-        title=TOPIC1_TITLE,
-        lecture=lecture1,
-        type=TOPIC1_TYPE,
-        description=TOPIC1_DESCRIPTION,
-        order=1,
-        is_premium=True,
-    )
-    topic2 = Topic.objects.create(
-        title=TOPIC2_TITLE,
-        lecture=lecture2,
-        type=TOPIC2_TYPE,
-        description=TOPIC2_DESCRIPTION,
-        order=TOPIC2_ORDER,
-        is_premium=True,
-    )
-    assignment = Assignment.objects.create(question=ASSIGNMENT_QUESTION, topic=topic1)
-    mcq = MultipleChoiceQuestion.objects.create(
-        question=MCQ_QUESTION,
-        topic=topic2,
-    )
-    choice1 = MultipleChoiceQuestionChoice.objects.create(
-        choice=MCQ_CHOICE1,
-        is_correct=True,
-        question=mcq,
-    )
-    choice2 = MultipleChoiceQuestionChoice.objects.create(
-        choice=MCQ_CHOICE2,
-        is_correct=False,
-        question=mcq,
-    )
-    choice3 = MultipleChoiceQuestionChoice.objects.create(
-        choice=MCQ_CHOICE3,
-        is_correct=False,
-        question=mcq,
-    )
-    choice4 = MultipleChoiceQuestionChoice.objects.create(
-        choice=MCQ_CHOICE4,
-        is_correct=False,
-        question=mcq,
-    )
-    return {
-        "course": course,
-        "lecture1": lecture1,
-        "lecture2": lecture2,
-        "topic1": topic1,
-        "topic2": topic2,
-        "assignment": assignment,
-        "mcq": mcq,
-        "choice1": choice1,
-        "choice2": choice2,
-        "choice3": choice3,
-        "choice4": choice4,
-    }
+from courses.test import conftest
 
 
 @pytest.mark.django_db
@@ -139,34 +28,40 @@ class TestCourseDetailSerializer:
 
         # Then
         assert data["id"] == self.course.id
-        assert data["title"] == COURSE_TITLE
-        assert data["short_description"] == COURSE_SHORT_DESCRIPTION
-        assert data["description"] == COURSE_DESCRIPTION
-        assert data["category"] == COURSE_CATEGORY
-        assert data["course_level"] == COURSE_COURSE_LEVEL
+        assert data["title"] == conftest.COURSE_TITLE
+        assert data["short_description"] == conftest.COURSE_SHORT_DESCRIPTION
+        assert data["description"] == conftest.COURSE_DESCRIPTION
+        assert data["category"] == conftest.COURSE_CATEGORY
+        assert data["course_level"] == conftest.COURSE_COURSE_LEVEL
         assert len(data["lectures"]) == 2
-        assert data["lectures"][0]["title"] == LECTURE1_TITLE
-        assert data["lectures"][0]["order"] == LECTURE1_ORDER
-        assert data["lectures"][1]["title"] == LECTURE2_TITLE
-        assert data["lectures"][1]["order"] == LECTURE2_ORDER
+        assert data["lectures"][0]["title"] == conftest.LECTURE1_TITLE
+        assert data["lectures"][0]["order"] == conftest.LECTURE1_ORDER
+        assert data["lectures"][1]["title"] == conftest.LECTURE2_TITLE
+        assert data["lectures"][1]["order"] == conftest.LECTURE2_ORDER
         assert len(data["lectures"][0]["topics"]) == 1
-        assert data["lectures"][0]["topics"][0]["title"] == TOPIC1_TITLE
-        assert data["lectures"][0]["topics"][0]["type"] == TOPIC1_TYPE
-        assert data["lectures"][0]["topics"][0]["description"] == TOPIC1_DESCRIPTION
-        assert data["lectures"][0]["topics"][0]["order"] == TOPIC1_ORDER
+        assert data["lectures"][0]["topics"][0]["title"] == conftest.TOPIC1_TITLE
+        assert data["lectures"][0]["topics"][0]["type"] == conftest.TOPIC1_TYPE
+        assert (
+            data["lectures"][0]["topics"][0]["description"]
+            == conftest.TOPIC1_DESCRIPTION
+        )
+        assert data["lectures"][0]["topics"][0]["order"] == conftest.TOPIC1_ORDER
         assert data["lectures"][0]["topics"][0]["is_premium"] is True
         assert (
             data["lectures"][0]["topics"][0]["assignment"]["question"]
-            == ASSIGNMENT_QUESTION
+            == conftest.ASSIGNMENT_QUESTION
         )
-        assert data["lectures"][1]["topics"][0]["title"] == TOPIC2_TITLE
-        assert data["lectures"][1]["topics"][0]["type"] == TOPIC2_TYPE
-        assert data["lectures"][1]["topics"][0]["description"] == TOPIC2_DESCRIPTION
-        assert data["lectures"][1]["topics"][0]["order"] == TOPIC2_ORDER
+        assert data["lectures"][1]["topics"][0]["title"] == conftest.TOPIC2_TITLE
+        assert data["lectures"][1]["topics"][0]["type"] == conftest.TOPIC2_TYPE
+        assert (
+            data["lectures"][1]["topics"][0]["description"]
+            == conftest.TOPIC2_DESCRIPTION
+        )
+        assert data["lectures"][1]["topics"][0]["order"] == conftest.TOPIC2_ORDER
         assert data["lectures"][1]["topics"][0]["is_premium"] is True
         assert (
             data["lectures"][1]["topics"][0]["multiple_choice_question"]["question"]
-            == MCQ_QUESTION
+            == conftest.MCQ_QUESTION
         )
         assert (
             len(
@@ -180,7 +75,7 @@ class TestCourseDetailSerializer:
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][0]["choice"]
-            == MCQ_CHOICE1
+            == conftest.MCQ_CHOICE1
         )
         assert (
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
@@ -192,7 +87,7 @@ class TestCourseDetailSerializer:
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][1]["choice"]
-            == MCQ_CHOICE2
+            == conftest.MCQ_CHOICE2
         )
         assert (
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
@@ -204,7 +99,7 @@ class TestCourseDetailSerializer:
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][2]["choice"]
-            == MCQ_CHOICE3
+            == conftest.MCQ_CHOICE3
         )
         assert (
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
@@ -216,7 +111,7 @@ class TestCourseDetailSerializer:
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][3]["choice"]
-            == MCQ_CHOICE4
+            == conftest.MCQ_CHOICE4
         )
         assert (
             data["lectures"][1]["topics"][0]["multiple_choice_question"][
@@ -293,10 +188,10 @@ class TestCourseSummarySerializer:
 
         # Then
         assert data["id"] == course.id
-        assert data["title"] == COURSE_TITLE
-        assert data["short_description"] == COURSE_SHORT_DESCRIPTION
-        assert data["category"] == COURSE_CATEGORY
-        assert data["course_level"] == COURSE_COURSE_LEVEL
+        assert data["title"] == conftest.COURSE_TITLE
+        assert data["short_description"] == conftest.COURSE_SHORT_DESCRIPTION
+        assert data["category"] == conftest.COURSE_CATEGORY
+        assert data["course_level"] == conftest.COURSE_COURSE_LEVEL
 
     def test_course_역직렬화(self):
         # Given
@@ -328,15 +223,18 @@ class TestLectureSerializer:
 
         # Then
         assert data["id"] == self.lecture.id
-        assert data["title"] == LECTURE2_TITLE
+        assert data["title"] == conftest.LECTURE2_TITLE
         assert data["order"] == 2
         assert len(data["topics"]) == 1
-        assert data["topics"][0]["title"] == TOPIC2_TITLE
-        assert data["topics"][0]["type"] == TOPIC2_TYPE
-        assert data["topics"][0]["description"] == TOPIC2_DESCRIPTION
-        assert data["topics"][0]["order"] == TOPIC2_ORDER
+        assert data["topics"][0]["title"] == conftest.TOPIC2_TITLE
+        assert data["topics"][0]["type"] == conftest.TOPIC2_TYPE
+        assert data["topics"][0]["description"] == conftest.TOPIC2_DESCRIPTION
+        assert data["topics"][0]["order"] == conftest.TOPIC2_ORDER
         assert data["topics"][0]["is_premium"] is True
-        assert data["topics"][0]["multiple_choice_question"]["question"] == MCQ_QUESTION
+        assert (
+            data["topics"][0]["multiple_choice_question"]["question"]
+            == conftest.MCQ_QUESTION
+        )
         assert (
             len(
                 data["topics"][0]["multiple_choice_question"][
@@ -349,7 +247,7 @@ class TestLectureSerializer:
             data["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][0]["choice"]
-            == MCQ_CHOICE1
+            == conftest.MCQ_CHOICE1
         )
         assert (
             data["topics"][0]["multiple_choice_question"][
@@ -361,7 +259,7 @@ class TestLectureSerializer:
             data["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][1]["choice"]
-            == MCQ_CHOICE2
+            == conftest.MCQ_CHOICE2
         )
         assert (
             data["topics"][0]["multiple_choice_question"][
@@ -373,7 +271,7 @@ class TestLectureSerializer:
             data["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][2]["choice"]
-            == MCQ_CHOICE3
+            == conftest.MCQ_CHOICE3
         )
         assert (
             data["topics"][0]["multiple_choice_question"][
@@ -385,7 +283,7 @@ class TestLectureSerializer:
             data["topics"][0]["multiple_choice_question"][
                 "multiple_choice_question_choices"
             ][3]["choice"]
-            == MCQ_CHOICE4
+            == conftest.MCQ_CHOICE4
         )
         assert (
             data["topics"][0]["multiple_choice_question"][
@@ -441,12 +339,12 @@ class TestTopicSerializer:
 
         # Then
         assert data["id"] == self.topic.id
-        assert data["title"] == TOPIC2_TITLE
-        assert data["type"] == TOPIC2_TYPE
-        assert data["description"] == TOPIC2_DESCRIPTION
-        assert data["order"] == TOPIC2_ORDER
+        assert data["title"] == conftest.TOPIC2_TITLE
+        assert data["type"] == conftest.TOPIC2_TYPE
+        assert data["description"] == conftest.TOPIC2_DESCRIPTION
+        assert data["order"] == conftest.TOPIC2_ORDER
         assert data["is_premium"] is True
-        assert data["multiple_choice_question"]["question"] == MCQ_QUESTION
+        assert data["multiple_choice_question"]["question"] == conftest.MCQ_QUESTION
         assert (
             len(data["multiple_choice_question"]["multiple_choice_question_choices"])
             == 4
@@ -455,7 +353,7 @@ class TestTopicSerializer:
             data["multiple_choice_question"]["multiple_choice_question_choices"][0][
                 "choice"
             ]
-            == MCQ_CHOICE1
+            == conftest.MCQ_CHOICE1
         )
         assert (
             data["multiple_choice_question"]["multiple_choice_question_choices"][0][
@@ -467,7 +365,7 @@ class TestTopicSerializer:
             data["multiple_choice_question"]["multiple_choice_question_choices"][1][
                 "choice"
             ]
-            == MCQ_CHOICE2
+            == conftest.MCQ_CHOICE2
         )
         assert (
             data["multiple_choice_question"]["multiple_choice_question_choices"][1][
@@ -479,7 +377,7 @@ class TestTopicSerializer:
             data["multiple_choice_question"]["multiple_choice_question_choices"][2][
                 "choice"
             ]
-            == MCQ_CHOICE3
+            == conftest.MCQ_CHOICE3
         )
         assert (
             data["multiple_choice_question"]["multiple_choice_question_choices"][2][
@@ -491,7 +389,7 @@ class TestTopicSerializer:
             data["multiple_choice_question"]["multiple_choice_question_choices"][3][
                 "choice"
             ]
-            == MCQ_CHOICE4
+            == conftest.MCQ_CHOICE4
         )
         assert (
             data["multiple_choice_question"]["multiple_choice_question_choices"][3][
@@ -540,7 +438,7 @@ class TestAssignmentSerializer:
 
         # Then
         assert data["id"] == assignment.id
-        assert data["question"] == "Test Assignment"
+        assert data["question"] == conftest.ASSIGNMENT_QUESTION
 
     def test_assignment_역직렬화(self):
         # Given
@@ -596,7 +494,7 @@ class TestMultipleChoiceQuestionSerializer:
 
         # Then
         assert data["id"] == self.mcq.id
-        assert data["question"] == "Test Multiple Choice Question"
+        assert data["question"] == conftest.MCQ_QUESTION
         assert len(data["multiple_choice_question_choices"]) == 4
         assert data["multiple_choice_question_choices"][0]["choice"] == "Choice 1"
         assert data["multiple_choice_question_choices"][0]["is_correct"] is True
